requirements.txt
faiss-cpu
numpy
psycopg[binary]
python-dotenv
pgvector
sentence-transformers
pymupdf
pytest
